#434: Moved duplicated method into super class SkillModifierIncreaseEffect
 Now it seems to work.
